Update Login2Fa
- Update Login2Fa to know have a endpoint using FastEndpoints.
- Add Login2Fa method to AuthService
