+++ src/Application/Eiromplays.IdentityServer.Application/Identity/Auth/IAuthService.cs
@@ -9,4 +9,6 @@ public interface IAuthService : ITransientService
     Task<dynamic> LogoutAsync(LogoutRequest request, HttpContext httpContext);
 
     Task<Result<LoginResponse>> LoginAsync(LoginRequest request);
+
+    Task<Result<LoginResponse>> Login2FaAsync(Login2FaRequest request);
 }
