+++ src/Application/Eiromplays.IdentityServer.Application/Identity/Auth/Requests/Login/LoginRequest.cs
@@ -9,3 +9,20 @@ public class LoginRequest
     
     public string? ReturnUrl { get; set; }
 }
+
+public class LoginRequestValidator : Validator<LoginRequest>
+{
+    public LoginRequestValidator()
+    {
+        RuleFor(x => x.Login)
+            .NotEmpty()
+            .MaximumLength(100);
+        
+        RuleFor(x => x.Password)
+            .NotEmpty()
+            .MaximumLength(100);
+
+        RuleFor(x => x.ReturnUrl)
+            .MaximumLength(2000);
+    }
+}
