+++ src/Presentation/Eiromplays.IdentityServer/Endpoints/v1/Account/Login2FaEndpoint.cs
@@ -1,27 +1,17 @@
-using System.Net;
-using Duende.IdentityServer.Services;
-using Eiromplays.IdentityServer.Application.Common.Interfaces;
-using Eiromplays.IdentityServer.Contracts.v1.Requests.Account;
-using Eiromplays.IdentityServer.Contracts.v1.Responses.Account;
-using Eiromplays.IdentityServer.Infrastructure.Identity.Entities;
-using Microsoft.AspNetCore.Identity;
+using Eiromplays.IdentityServer.Application.Common.Exceptions;
+using Eiromplays.IdentityServer.Application.Identity.Auth;
+using Eiromplays.IdentityServer.Application.Identity.Auth.Requests.Login;
+using Eiromplays.IdentityServer.Application.Identity.Auth.Responses.Login;
 
 namespace Eiromplays.IdentityServer.Endpoints.v1.Account;
 
 public class Login2FaEndpoint : Endpoint<Login2FaRequest, LoginResponse>
 {
-    private readonly IUserResolver<ApplicationUser> _userResolver;
-    private readonly SignInManager<ApplicationUser> _signInManager;
-    private readonly IIdentityServerInteractionService _interaction;
-    private readonly IServerUrls _serverUrls;
+    private readonly IAuthService _authService;
 
-    public Login2FaEndpoint(IUserResolver<ApplicationUser> userResolver, SignInManager<ApplicationUser> signInManager,
-        IIdentityServerInteractionService interaction, IServerUrls serverUrls)
+    public Login2FaEndpoint(IAuthService authService)
     {
-        _userResolver = userResolver;
-        _signInManager = signInManager;
-        _interaction = interaction;
-        _serverUrls = serverUrls;
+        _authService = authService;
     }
 
     public override void Configure()
@@ -36,26 +26,16 @@ public class Login2FaEndpoint : Endpoint<Login2FaRequest, LoginResponse>
     
     public override async Task HandleAsync(Login2FaRequest req, CancellationToken ct)
     {
-        var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
-        if (user is null)
-            throw new InvalidOperationException("Unable to get user");
-            
-        var authenticatorCode = req.TwoFactorCode?.Replace(" ", string.Empty).Replace("-", string.Empty);
+        var result = await _authService.Login2FaAsync(req);
 
-        var result =
-            await _signInManager.TwoFactorAuthenticatorSignInAsync(authenticatorCode, req.RememberMe, req.RememberMachine);
-
-        if (result.Succeeded)
+        await result.Match(async res => await SendOkAsync(res, ct), exception =>
         {
-            var url = !string.IsNullOrWhiteSpace(req.ReturnUrl) ? Uri.UnescapeDataString(req.ReturnUrl) : null;
-            
-            if (_interaction.IsValidReturnUrl(url))
-                Response.ValidReturnUrl = url ?? _serverUrls.BaseUrl;
+            if (exception is BadRequestException badRequestException)
+            {
+                ThrowError(badRequestException.Message);
+            }
 
-            await SendOkAsync(Response, ct);
-            return;
-        }
-
-        ThrowError("Invalid authentication code");
+            return SendErrorsAsync(cancellation: ct);
+        });
     }
 }
