--- src/Presentation/Eiromplays.IdentityServer/Extensions/Extensions.cs
@@ -1,26 +1,0 @@
-// Copyright (c) Duende Software. All rights reserved.
-// See LICENSE in the project root for license information.
-
-// Original file: https://github.com/DuendeSoftware/Samples/blob/main/IdentityServer/v6/Quickstarts
-// Modified by Eirik Sjøløkken
-
-using Duende.IdentityServer.Models;
-using Eiromplays.IdentityServer.ViewModels.Account;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Eiromplays.IdentityServer.Extensions
-{
-    public static class Extensions
-    {
-        /// <summary>
-        /// Checks if the redirect URI is for a native client.
-        /// </summary>
-        /// <returns></returns>
-        public static bool IsNativeClient(this AuthorizationRequest context)
-        {
-            return !context.RedirectUri.StartsWith("https", StringComparison.Ordinal)
-                && !context.RedirectUri.StartsWith("http", StringComparison.Ordinal);
-        }
-    }
-}
