--- src/Presentation/Eiromplays.IdentityServer/ViewModels/Account/ConfirmEmailViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Eiromplays.IdentityServer.ViewModels.Account;
-
-public class ConfirmEmailViewModel
-{
-    public string? Username { get; set; }
-    public string? VerificationUrl { get; set; }
-}
