--- src/Presentation/Eiromplays.IdentityServer/ViewModels/Account/LoggedOutViewModel.cs
@@ -1,20 +1,0 @@
-﻿// Copyright (c) Duende Software. All rights reserved.
-// See LICENSE in the project root for license information.
-
-// Original file: https://github.com/DuendeSoftware/Samples/blob/main/IdentityServer/v6/Quickstarts
-// Modified by Eirik Sjøløkken
-
-namespace Eiromplays.IdentityServer.ViewModels.Account;
-
-public class LoggedOutViewModel
-{
-    public string? PostLogoutRedirectUri { get; set; }
-    public string? ClientName { get; set; }
-    public string? SignOutIframeUrl { get; set; }
-
-    public bool AutomaticRedirectAfterSignOut { get; set; }
-
-    public string? LogoutId { get; set; }
-    public bool TriggerExternalSignout => ExternalAuthenticationScheme is not null;
-    public string? ExternalAuthenticationScheme { get; set; }
-}
