--- src/Presentation/Eiromplays.IdentityServer/ViewModels/Account/LoginViewModel.cs
@@ -1,27 +1,0 @@
-﻿// Copyright (c) Duende Software. All rights reserved.
-// See LICENSE in the project root for license information.
-
-// Original file: https://github.com/DuendeSoftware/Samples/blob/main/IdentityServer/v6/Quickstarts
-// Modified by Eirik Sjøløkken
-
-using System.ComponentModel.DataAnnotations;
-
-namespace Eiromplays.IdentityServer.ViewModels.Account;
-
-public class LoginViewModel : LoginInputModel
-{
-    public bool AllowRememberLogin { get; set; } = true;
-    public bool EnableLocalLogin { get; set; } = true;
-
-    public IEnumerable<ExternalProvider> ExternalProviders { get; set; } = Enumerable.Empty<ExternalProvider>();
-    public IEnumerable<ExternalProvider> VisibleExternalProviders => ExternalProviders.Where(x => !string.IsNullOrWhiteSpace(x.DisplayName));
-
-    public bool IsExternalLoginOnly => !EnableLocalLogin && ExternalProviders.Count() == 1;
-    public string? ExternalLoginScheme => IsExternalLoginOnly ? ExternalProviders.SingleOrDefault()?.AuthenticationScheme : null;
-
-    [Required] public string Login { get; set; } = default!;
-
-    [Required] public string Password { get; set; } = default!;
-    public bool RememberLogin { get; set; }
-    public string? ReturnUrl { get; set; }
-}
