--- src/Presentation/Eiromplays.IdentityServer/ViewModels/Account/RegisterWithoutUsernameViewModel.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Eiromplays.IdentityServer.ViewModels.Account;
-
-public class RegisterWithoutUsernameViewModel
-{
-    [Required]
-    [EmailAddress]
-    public string? Email { get; set; }
-
-    [Required]
-    [DataType(DataType.Password)]
-    public string? Password { get; set; }
-
-    [DataType(DataType.Password)]
-    [Compare("Password")]
-    public string? ConfirmPassword { get; set; }
-}
