--- src/Presentation/Eiromplays.IdentityServer/ViewModels/Account/ResendEmailConfirmationViewModel.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Eiromplays.IdentityServer.ViewModels.Account;
-
-public class ResendEmailConfirmationViewModel
-{
-    [Required]
-    public string? Identifier { get; set; }
-}
